# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: API
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /subscriptions/me:
    get:
      operationId: getMySubscription
      summary: Get current user subscription
      tags: [subscriptions]
      responses:
        "200":
          description: Subscription info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "404":
          description: No subscription found
  /subscriptions:
    get:
      operationId: getAllSubscriptions
      summary: Admin — list all subscriptions
      tags: [subscriptions]
      responses:
        "200":
          description: All subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      summary: Create subscription for a user
      tags: [subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    patch:
      operationId: updateSubscription
      summary: Admin — update subscription status or plan
      tags: [subscriptions]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionRequest"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /plans:
    get:
      operationId: getPlans
      summary: List active plans (public)
      tags: [plans]
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [client, broker, correspondent] }
        - name: includeInactive
          in: query
          schema: { type: string }
        - name: includeLegacy
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Plans list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      summary: Admin — create a new plan
      tags: [plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      summary: Get a single plan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
        "404":
          description: Not found
    patch:
      operationId: updatePlan
      summary: Admin — update a plan
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      summary: Admin — deactivate a plan (soft delete)
      tags: [plans]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
...
components:
  schemas:
...
    Subscription:
      type: object
      required: [id, userId, userName, userEmail, userRole, plan, status, priceMonthly, billingDay, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userRole:
          type: string
        plan:
          type: string
          enum: [free, individual, plus, corretor, imobiliaria, enterprise, correspondente_individual, correspondente_sucesso, bank_connect, corretor_50, corretor_200, corretor_enterprise, correspondent_50, correspondent_200, correspondent_enterprise, client, correspondent]
        status:
          type: string
          enum: [trial, active, overdue, cancelled, inactive]
        priceMonthly:
          type: number
        billingDay:
          type: integer
        marketplaceAddon:
          type: boolean
          nullable: true
        marketplacePropertyLimit:
          type: integer
          nullable: true
        marketplaceAddonPrice:
          type: number
          nullable: true
        trialEndsAt:
          type: string
          format: date-time
          nullable: true
        lastPaymentAt:
          type: string
          format: date-time
          nullable: true
        nextDueAt:
          type: string
          format: date-time
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSubscriptionRequest:
      type: object
      required: [userId, userName, userEmail, userRole, plan]
      properties:
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        userRole:
          type: string
        plan:
          type: string
          enum: [free, individual, plus, corretor, imobiliaria, enterprise, correspondente_individual, correspondente_sucesso, bank_connect, corretor_50, corretor_200, corretor_enterprise, correspondent_50, correspondent_200, correspondent_enterprise, client, correspondent]
        status:
          type: string
          enum: [trial, active, overdue, cancelled, inactive]
        billingDay:
          type: integer
        marketplaceAddon:
          type: boolean
        marketplacePropertyLimit:
          type: integer
        marketplaceAddonPrice:
          type: number
        notes:
          type: string
    UpdateSubscriptionRequest:
      type: object
      properties:
        plan:
          type: string
          enum: [free, individual, plus, corretor, imobiliaria, enterprise, correspondente_individual, correspondente_sucesso, bank_connect, corretor_50, corretor_200, corretor_enterprise, correspondent_50, correspondent_200, correspondent_enterprise, client, correspondent]
        status:
          type: string
          enum: [trial, active, overdue, cancelled, inactive]
        billingDay:
          type: integer
        marketplaceAddon:
          type: boolean
        marketplacePropertyLimit:
          type: integer
        marketplaceAddonPrice:
          type: number
        lastPaymentAt:
          type: string
          format: date-time
        nextDueAt:
          type: string
          format: date-time
        notes:
          type: string
...
    Plan:
      type: object
      required: [id, label, role, group, priceMonthly, priceYearly, highlight, enterprise, color, bgLight, description, features, isActive, isLegacy, sortOrder]
      properties:
        id: { type: string }
        label: { type: string }
        role:
          type: string
          enum: [client, broker, correspondent]
        group:
          type: string
          enum: [individual, corretor, correspondent]
        priceMonthly: { type: number }
        priceYearly: { type: number }
        highlight: { type: boolean }
        leadLimit: { type: integer, nullable: true }
        userLimit: { type: integer, nullable: true }
        enterprise: { type: boolean }
        color: { type: string }
        bgLight: { type: string }
        description: { type: string }
        features:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        isLegacy: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreatePlanRequest:
      type: object
      required: [id, label, role, group, priceMonthly]
      properties:
        id: { type: string }
        label: { type: string }
        role:
          type: string
          enum: [client, broker, correspondent]
        group:
          type: string
          enum: [individual, corretor, correspondent]
        priceMonthly: { type: number }
        priceYearly: { type: number }
        highlight: { type: boolean }
        leadLimit: { type: integer, nullable: true }
        userLimit: { type: integer, nullable: true }
        enterprise: { type: boolean }
        color: { type: string }
        bgLight: { type: string }
        description: { type: string }
        features:
          type: array
          items: { type: string }
        sortOrder: { type: integer }
    UpdatePlanRequest:
      type: object
      properties:
        label: { type: string }
        priceMonthly: { type: number }
        priceYearly: { type: number }
        highlight: { type: boolean }
        leadLimit: { type: integer, nullable: true }
        userLimit: { type: integer, nullable: true }
        enterprise: { type: boolean }
        color: { type: string }
        bgLight: { type: string }
        description: { type: string }
        features:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }