# lib/api-zod/src/generated/types/createPlanRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { CreatePlanRequestGroup } from "./createPlanRequestGroup";
import type { CreatePlanRequestRole } from "./createPlanRequestRole";

export interface CreatePlanRequest {
  id: string;
  label: string;
  role: CreatePlanRequestRole;
  group: CreatePlanRequestGroup;
  priceMonthly: number;
  priceYearly?: number;
  highlight?: boolean;
  leadLimit?: number | null;
  enterprise?: boolean;
  color?: string;
  bgLight?: string;
  description?: string;
  features?: string[];
  sortOrder?: number;
}